+++ src/classifiers.js
@@ -4,21 +4,20 @@
  * @module classifiers
  */
 
-const {hasComponentTag, hasHOCTag} = require('./utils');
+import {hasComponentTag, hasHOCTag} from './utils.js';
 
-function defaultTypeClassifier ({section}) {
+export async function defaultTypeClassifier ({section}) {
 	const kind = section.kind;
 
 	// Check for HOC
-	if (hasHOCTag(section)) {
+	if (await hasHOCTag(section)) {
 		return 'hoc';
 	}
 
 	// Check for @ui
-	if (hasComponentTag(section)) {
+	if (await hasComponentTag(section)) {
 		return 'component';
 	}
 
 	return kind;
 }
-exports.defaultTypeClassifier = defaultTypeClassifier;