+++ src/utils.js
@@ -1,4 +1,4 @@
-const jsonata = require('jsonata');
+import jsonata from 'jsonata';
 
 /**
  * Checks member for a 'required' tag
@@ -6,24 +6,25 @@ const jsonata = require('jsonata');
  * @param {Object} member A property definition
  * @returns {Boolean}
  */
-const hasRequiredTag = (member) => {
+export const hasRequiredTag = async (member) => {
 	// Find any tag field whose `title` is 'required' (won't be there if not required)
 	const expression = "$[title='required']";
-	const result = jsonata(expression).evaluate(member.tags);
+	const result = await jsonata(expression).evaluate(member.tags);
 	return !!result;
 };
-exports.hasRequiredTag = hasRequiredTag;
+
 /**
  * Checks member for a HOC tag
+ *
  * @param {Object} member
+ * @returns {Boolean}
  */
-const hasHOCTag = (member) => {
+export const hasHOCTag = async (member) => {
 	// Find any tag field whose `title` is 'hoc'
 	const expression = "$[title='hoc']";
-	const result = jsonata(expression).evaluate(member.tags);
+	const result = await jsonata(expression).evaluate(member.tags);
 	return !!result;
 };
-exports.hasHOCTag = hasHOCTag;
 
 /**
  * Checks member for a 'ui' tag
@@ -31,16 +32,13 @@ exports.hasHOCTag = hasHOCTag;
  * @param {Object} member A property definition
  * @returns {Boolean}
  */
-const hasComponentTag = (member) => {
-	// Find any tag field whose `title` is 'required' (won't be there if not required)
+export const hasComponentTag = async (member) => {
+	// Find any tag field whose `title` is 'ui' (won't be there if not required)
 	const expression = "$[title='ui']";
-	const result = jsonata(expression).evaluate(member.tags);
+	const result = await jsonata(expression).evaluate(member.tags);
 	return !!result;
 };
-exports.hasComponentTag = hasComponentTag;
 
-const escapeClassMember = (member) => {
+export const escapeClassMember = (member) => {
 	return member.match(/-/) ? `'${member}'` : member;
 };
-
-exports.escapeClassMember = escapeClassMember;